DF-9778 win10 fix
Former-commit-id: 420b1a2935851e165388f48b34eb0758b245a067

+++ Sources/Internal/Utils/Utils.cpp
@@ -123,7 +123,7 @@ int read_handler(void* ext, unsigned char* buffer, size_t size, size_t* length)
 
     *length = sizeToWrite;
 
-    holder->dataOffset += sizeToWrite;
+    holder->dataOffset += static_cast<uint32>(sizeToWrite);
 
     return 1;
 }
